% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat_preprocessing.R
\name{br_temp}
\alias{br_temp}
\title{br_temp}
\usage{
br_temp(
  x,
  band = "",
  conv = TRUE,
  mult = NULL,
  add = NULL,
  k1 = NULL,
  k2 = NULL
)
}
\arguments{
\item{x}{Image band in DN to be converted.}

\item{band}{Character. Number of the Landsat band to convert.}

\item{conv}{Logical; if TRUE, units are converted to Celsius degrees.}

\item{mult}{Radiance multiplicative band rescaling factor.}

\item{add}{Radiance additive band rescaling factor.}

\item{k1}{k1 thermal conversion constant.}

\item{k2}{k2 thermal conversion constant.}
}
\value{
Raster layer with TOA brightness temperature values
in Kelvin or Celsius degrees.
}
\description{
Convert the DN contained in a Landsat TIR band
to TOA brightness temperature.
}
\details{
Convert the DN of a Landsat TIR band to TOA brightness
temperature in Kelvin or Celsius degrees employing the radiance
multiplicative and additive band rescaling factors and K1 and K2
constants. If band is specified, the function reads the
metadata (.txt) directly from the work directory (folder
containing bands as downloaded from NASA EarthExplorer) and
automatically extracts the multiplicative and additive rescaling
factors and k1 and k2 constants. These scaling factors and
constants can be manually defined employing mult, add, k1 and
k2 parameters. In this case band is ignored. If parameter
conv = TRUE, temperature units are convert to Celsius degrees.
This is the default.
}
\examples{
# For Landsat 8 band 10 defining band and extracting scaling
factors and constants from metadata
brtempB10 <- br_temp(B10, band = "10", conv = TRUE)
# For Landsat 8 band 10 defining manually the multiplicative
and additive scaling factors and the k1 and k2 constants
brtempB10 <- br_temp(B10, conv = TRUE, mult = 0.00033420,
add = 0.1, k1 = 774.8853, k2 = 1321.0789)

}
\references{
USGS. (2019). Landsat 8 data users handbook version 4.
USGS Earth Resources Observation and Science (EROS). Sioux Falls,
South Dakota. USA. 106.
}
