% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat_processing.R
\name{emissivity}
\alias{emissivity}
\title{emissivity}
\usage{
emissivity(
  x,
  nonveg = 0.2,
  veg = 0.5,
  enonveg = 0.95,
  eveg = 0.99,
  pveg = FALSE
)
}
\arguments{
\item{x}{NDVI raster layer.}

\item{nonveg}{NDVI threshold value for non-vegetated pixels.}

\item{veg}{NDVI threshold value for vegetated pixels.}

\item{enonveg}{Emissivity value for non-vegetated pixels.}

\item{eveg}{Emissivity value for vegetated pixels.}

\item{pveg}{Logical; if TRUE, proportion of vegetation is given as an output.}
}
\value{
Raster layer containing emissivity values. If pveg = TRUE,
a list with emissivity and pveg raster layers.
}
\description{
Compute land surface emissivity.
}
\details{
Computes the per-pixel emissivity based on the modified
NDVI threshold method (Sobrino et al., 2008). Threshold values
for non-vegetated and vegetated pixels can be defined by the user.
Defaults are nonveg <= 0.2 and veg >= 0.5, based on the NDVI values.
Emissivity values for non-vegetated and vegetated pixels can be
also defined. Defaults are enonveg = 0.95 and eveg = 0.99.  pveg
is logical. If TRUE, the proportion of vegetation layer is added
to the output. Default is FALSE.
}
\examples{
# For default non-vegetation and vegetation emissivity values
emiss <- emissivity(ndviind, enonveg = 0.95, eveg = 0.99, pveg = TRUE)

}
\references{
Sobrino, J. A., Jim?nez-Mu?oz, J. C., S?ria, G.,
Romaguera, M., Guanter, L., Moreno, J. & Mart?nez, P. (2008).
Land surface emissivity retrieval from different VNIR and TIR
sensors. IEEE Transactions on Geoscience and Remote Sensing,
46(2): 316-327.
}
