% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UHI_indicators.R
\name{getis}
\alias{getis}
\title{getis}
\usage{
getis(x, dist = NULL, p = 0.05)
}
\arguments{
\item{x}{Raster layer with values to analyze.}

\item{dist}{Distance for local neighborhood size from each cell.}

\item{p}{P-value to reject the null hypothesis.}
}
\value{
A shapefile with the resulting z-scores, FDR values
and label of cluster (i.e., hot spot, cold spot or no sig).
}
\description{
Compute a hot and cold spots analysis employing
Getis-Ord Gi* statistic and the False Discovery Rate (FDR)
correction.
}
\details{
Compute a hot and cold spots analysis based on the
Getis-Ord Gi* statistic and the False Discovery Rate (FDR)
correction. Center cell is included in the local neighborhood
size and the limit is the distance specified by dist employing
queen's case. False Discovery Rate (FDR) correction is applied
to potentially reduce the critical p-value thresholds in order
to account for multiple testing and spatial dependency. Possible
p values to reject the null hypothesis are 0.1, 0.05, 0.01,
0.001 and 0.0001.
}
\examples{
\dontrun{# For an LST layer derived from Landsat 8
spots <- getis(lstl8, dist = 70, p = 0.05)}

}
\references{
Benjamini, Y., and Hochberg, Y. (1995). Controlling
the false discovery rate: a practical and powerful approach
to multiple testing. Journal of the Royal Statistical Society
Series B, 57, 289-300.

Getis, A. and Ord, J. K. (1996). Local spatial statistics:
an overview. In P. Longley and M. Batty (eds) Spatial analysis:
modelling in a GIS environment (Cambridge: Geoinformation
International), 261-277.
}
