% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modis_preprocessing.R
\name{hdf_extract}
\alias{hdf_extract}
\title{hdf_extract}
\usage{
hdf_extract(x, dest = getwd(), conv = TRUE)
}
\arguments{
\item{x}{.hdf MODIS file.}

\item{dest}{Path or folder to export the extracted bands.}

\item{conv}{Logical; if TRUE, temperature units are converted to Celsius degrees.}
}
\value{
Exported .tif layers.
}
\description{
Extract and scale MODIS layers from a .hdf MODIS LST or
vegetation index product as downloaded from NASA EarthExplorer.
}
\details{
Extract MODIS layers from a .hdf MODIS product retaining
the product name, acquisition date, tile identifier, collection
version and julian date of production in the output name. The
extraction of names is suited specifically for vegetation indices
(MXD13) and LST and Emissivity (MXD11) MODIS products as downloaded
from NASA EarthExplorer. Products MXD11_L2 and MXD21_L2 are not
supported. The function scales the values of each band. If parameter
conv = TRUE, temperature units are converted to Celsius degrees for
LST bands. This is the default.

A negative sign of the viewing angle in MXD11 products means MODIS
viewing the grid from east. The view zenith angle itself is always
a positive number, the zenith angle from nadir. The information of
MODIS viewing the grid from east or west may be important in
understanding the view angle effect in the temporal variations in
LST, especially in rugged regions.
}
\examples{
# For MOD11A2 LST product saving .tif bands in a different folder
hdf_extract("MOD11A2.A2019169.h08v07.006.2019178033848.hdf",
"C:/MODISLST", conv = TRUE)

# For MOD13A3 product saing .tif bands in workdirectory
hdf_extract("MOD13A3.A2019121.h08v07.006.2019162004758.hdf")

}
\references{
https://earthexplorer.usgs.gov/

Wan, Z. (1999). "MODIS Land-Surface Temperature Algorithm
Theoretical Basis Document (LST ATBD) Version 3.3, April 1999".

Wolfe, R.E., D.P. Roy, E. Vermote. (1998). "MODIS land data storage,
gridding and compositing methodology: level 2 grid".
IEEE Trans. Geosci. Remote Sens., v36, n4, pp. 1324-1338.
}
