% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UHI_indicators.R
\name{hia}
\alias{hia}
\title{hia}
\usage{
hia(x, y = NULL)
}
\arguments{
\item{x}{LST raster layer.}

\item{y}{Optional. Shapefile polygon of the city extent.}
}
\value{
A list with the LST raster layer representing the
areas with values higher than the HIA threshold, LST mean,
LST sd, HIA threshold and HIA in hectares.
}
\description{
Computes the Heat Island Area (HIA) indicator based
on a LST raster layer.
}
\details{
Computes the Hot Island Area (HIA) indicator in hectares.
If y is supplied, it must have the same projection as x.
When y is not defined, it is assumed that the LST raster is
already cropped to the city extent.
}
\examples{
# Specifying city polygon
HIA.City <- hia(Llstl8, citypol)

}
\references{
Zhang, J. Q., and Wang, Y. P. (2008). Study of the
relationships between the spatial extent of surface urban heat
islands and urban characteristic factors based on Landsat ETM
plus data. Sensors, 8(11), 7453-7468.
}
