% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat_processing.R
\name{ibi}
\alias{ibi}
\title{ibi}
\usage{
ibi(x, y, r, g)
}
\arguments{
\item{x}{SWIR band.}

\item{y}{NIR band.}

\item{r}{Red band.}

\item{g}{Green band.}
}
\value{
IBI values ranging from -1 to 1.
}
\description{
Compute the Index based built-up Index (IBI).
}
\details{
Compute the Index based built-up Index (IBI) employing
the surface reflectance SWIR, NIR, Red and Green bands of a
satellite image.
}
\examples{
# For Landsat 8 bands
ibiind <- ibi(B6, B5, B4, B3)

}
\references{
Xu, H. (2008). A new index for delineating built-up
land features in satellite imagery. Int. J. Remote Sens. 29, 4269-4276.
}
