% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat_processing.R
\name{landsat_lst}
\alias{landsat_lst}
\title{landsat_lst}
\usage{
landsat_lst(x, y, sensor = NULL, conv = TRUE)
}
\arguments{
\item{x}{Brightness temperature layer.}

\item{y}{Emissivity layer.}

\item{sensor}{Landsat sensor of the input bands. Possible values are "L4", "L5", "L7" and "L8".}

\item{conv}{Logical; if TRUE, units are converted to Celsius degrees.}
}
\value{
LST layer in Celsius or Kelvin degrees.
}
\description{
Compute land surface temperature.
}
\details{
Compute the per-pixel LST of a Landsat brightness
temperature layer based on the inversion of the Planck function,
which corrects the bias present in brightness temperature values,
in which is assumed that the land surface is a black body with an
emissivity of 1.  Landsat sensor must be specified in order to
apply the proper center wavelength values for the corresponding
Landsat TIR band. If parameter conv = TRUE, temperature units
are convert from Kelvin to Celsius degrees. This is the default.
}
\examples{
# For LST derived from Landsat 8 sensor in Celsius degrees.
L8lst <- landsat_lst(Brtempb10, emiss, Sensor = "L8", conv = TRUE)

}
\references{
USGS. (2019). Landsat 7 data users handbook version 2.
USGS Earth Resources Observation and Science (EROS).
Sioux Falls, South Dakota. USA. 139.

USGS. (2019). Landsat 8 data users handbook version 4. USGS
Earth Resources Observation and Science (EROS). Sioux Falls,
South Dakota. USA. 106.

Stathopoulou, M. and Cartalis, C. (2007). Daytime urban heat
islands from Landsat ETM+ and Corine land cover data: An
application to major cities in Greece. Sol. Energy. 81 358-368.
}
