% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modis_preprocessing.R
\name{lst_filter}
\alias{lst_filter}
\title{lst_filter}
\usage{
lst_filter(x, time = "day", flag, angle, conv = TRUE)
}
\arguments{
\item{x}{.hdf MODIS file corresponding to a LST product.}

\item{time}{A character specifying the time of the data to filter.
Possible values are "day" or "night".}

\item{flag}{LST error bit flag or accuracy tolerance from 1-4.}

\item{angle}{Viewing zenith angle tolerance to filter pixels (in degrees).}

\item{conv}{Logical; if TRUE, units are converted to Celsius degrees.}
}
\value{
List of three raster layers: LST layer in Kelvin or
Celsius degrees, error or accuracy layer and viewing zenith
angle layer. Only pixels that are not NA in the original LST
layer are included in the error and view zenith angle layers.
}
\description{
Filter pixels of MODIS LST products based on the
LST error and angle layers.
}
\details{
Filter the pixels of a day or night data of a MODIS LST
product, based on tolerance values of LST error and zenith viewing
angle defined by the user. If parameter conv = TRUE, temperature
units are convert to Celsius degrees. This is the default. Products
MXD11_L2 and MYD21_L2 are not supported.

LST error bit flag to filter pixels must be defined as an integer
from 0-3 corresponding to: 0=00, 1=01, 2=10 and 3=11 in the 7 & 6
bit flags for 8 bit QC layers and 15 & 14 bit flags for 16 bit QC
layers. LST error flags differ between MXD11 and MXD21 products.
In MXD11 products, the flag represents average LST error values
in Kelvin degrees, with an increase of 1 degree. In MXD21 products,
the level of error is represented as accuracy in Kelvin degrees,
with increases of 0.5 degrees. See table 13 (MXD11; Wan, 2013) and
table 10 (MXD21; Hulley et al., 2016) in the MXD11 and MXD21 user
guides, respectively.

A negative sign of the viewing angle means MODIS viewing the grid
from east. The view zenith angle itself is always a positive number,
the zenith angle from nadir. The information of MODIS viewing
the grid from east or west may be important in understanding the
view angle effect in the temporal variations in LST, especially
in rugged regions.
}
\examples{
# For LST day MXD11 product, filtering pixels with LST error
<= 1?K and view zenith angle <= 35?
x <- "MOD11A1.A2019305.h14v09.006.2019306084028.hdf"
lstfiltered  <- lst_filter (x, time = "day", flag  = 1,
angle = 35, conv = TRUE)

}
\references{
Hulley, G., Freepartner, R., Malakar, N. & Sarkar, S.
(2016), Moderate Resolution Imaging Spectroradiometer (MODIS) - Land
Surface Temperature and Emissivity Product (MxD21) User Guide
Collection-6. Pasadena, EUA, 29.

Hulley, G., Hook, S., & Hughes, C. (2012). MODIS MxD21 Land
Surface Temperature and Emissivity Algorithm Theoretical Basis
Document. In: Jet Propulsion Laboratory, California Institute
of Technology, JPL Publication 12-17, August, 2012.

Z. Wan, & Z.-L. Li. (2011), Chapter 25, MODIS land surface temperature
and emissivity. In B. Ramachandran, C. O. Justice, & M. J.
Abrams (Eds.), Land remote sensing and global environmental
change, NASA's Earth observing system and the science of ASTER
and MODIS, NASA's Earth observing system and the science of ASTER
and MODIS. New York Dordrecht Heidelberg London: Springer,
http://dx.doi.org/10.1007/978-1-4419-6749-7.

Z. Wan. (2013), Collection-6 MODIS Land Surface Temperature Products
User's guide. Santa Barbara, USA, 33.

Z. Wan, Y. Zhang, Q. Zhang, and Z.-L. Li. (2002), "Validation of
the land-surface temperature products retrieved from Terra Moderate
Resolution Imaging Spectroradiometer data", Remote Sens.
Environ., 83, 163-180.

Gawuc, L., & Struzewska, J. (2016). Impact of MODIS quality
control on temporally aggregated urban surface temperature and
long-term surface urban heat island intensity. Remote Sensing,
8(5). https://doi.org/10.3390/rs8050374

Hu, L., Brunsell, N. A., Monaghan, A. J., Barlage, M.,
& Wilhelmi, O. V. (2014). How can we use MODIS land surface
temperature to validate long-term urban model simulations?
Journal of Geophysical Research, 119(6), 3185-3201.
https://doi.org/10.1002/2013JD021101
}
