% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat_processing.R
\name{msavi}
\alias{msavi}
\title{msavi}
\usage{
msavi(x, y)
}
\arguments{
\item{x}{NIR band.}

\item{y}{Red band.}
}
\value{
MSAVI values ranging from -1 to 1.
}
\description{
Compute the Modified Soil-Adjusted Vegetation Index (MSAVI).
}
\details{
Compute the Modified Soil-Adjusted Vegetation Index
(MSAVI) employing the surface reflectance NIR and Red bands
of a satellite image.
}
\examples{
\dontrun{# For Landsat 8 bands
msaviind <- msavi(B5, B4)}

}
\references{
Qi J., Kerr Y., Chehbouni A. (1994). External
factor consideration in vegetation index development.
Proc. of Physical Measurements and Signatures in Remote
Sensing, ISPRS, 723-730.
}
