% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat_processing.R
\name{ndbi}
\alias{ndbi}
\title{ndbi}
\usage{
ndbi(x, y)
}
\arguments{
\item{x}{SWIR band.}

\item{y}{NIR band.}
}
\value{
NDBI values ranging from -1 to 1.
}
\description{
Compute the Normalized Difference Built-up Index (NDBI).
}
\details{
Compute the Normalized Difference Built-up Index (NDBI)
employing the surface reflectance SWIR and NIR bands of a
satellite image.
}
\examples{
\dontrun{# For Landsat 8 bands
ndbiind <- ndbi(B6, B5)}

}
\references{
ZHA, Y., GAO, Y. and NI, S. (2003). Use of normalized
difference built-up index in automatically mapping urban areas
from TM imagery. International Journal of Remote Sensing, 24, pp.
583-594.
}
