% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat_processing.R
\name{ndvi}
\alias{ndvi}
\title{ndvi}
\usage{
ndvi(x, y)
}
\arguments{
\item{x}{NIR band.}

\item{y}{Red band.}
}
\value{
NDVI values ranging from -1 to 1.
}
\description{
Compute the Normalized Difference Vegetation Index (NDVI).
}
\details{
Compute the Normalized Difference Vegetation index (NDVI)
employing the surface reflectance NIR and Red bands of a satellite image.
}
\examples{
\dontrun{# For Landsat 8 bands
ndviind <- ndvi(B5, B4)}

}
\references{
Rouse, J.W., R.H. Haas, J.A. Schell, and D.W. Deering.
(1974). Monitoring vegetation systems in the Great Plains with
ERTS, In: S.C. Freden, E.P. Mercanti, and M. Becker (eds) Third
Earth Resources Technology Satellite-1 Syposium. Volume I:
Technical Presentations, NASA SP-351, NASA, Washington, D.C.,
pp. 309-317.
}
