% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat_processing.R
\name{savi}
\alias{savi}
\title{savi}
\usage{
savi(x, y)
}
\arguments{
\item{x}{NIR band.}

\item{y}{Red band.}
}
\value{
SAVI values ranging from -1 to 1.
}
\description{
Compute the Soil-Adjusted Vegetation Index (SAVI).
}
\details{
Compute the Soil-Adjusted Vegetation Index
(SAVI) employing the surface reflectance NIR and Red
bands of a satellite image.
}
\examples{
\dontrun{# For Landsat 8 bands
saviind <- savi(B5, B4)}

}
\references{
Huete, A.R. (1988). A soil-adjusted vegetation
index (SAVI). Remote Sens. Environ., 25, pp. 295-309.
}
