% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UHI_indicators.R
\name{space_time_trends}
\alias{space_time_trends}
\title{space_time_trends}
\usage{
space_time_trends(x, space = NULL, time = NULL, stat = "mean", dist, p)
}
\arguments{
\item{x}{Raster stack with at least 5 bands, each one
corresponding to one evenly distributed time-step with non-NA
values in valid cells.}

\item{space}{Optional. Number of cells for spatial aggregation.}

\item{time}{Optional. Number of time-steps for temporal aggregation.}

\item{stat}{Statistic employed for spatial and/or temporal
aggregation. Possible values are "mean" or "max". Default value
is "mean".}

\item{dist}{Distance for local neighborhood size from each cell.}

\item{p}{P-value to reject the null hypothesis.}
}
\value{
A shapefile containing the space-time trend ID, the
space-time trend label, the Mann-Kendall p-values, z-scores,
tau values and slope, the percentage of time-steps on which each
cell showed to be hot spot and the percentage of time-steps on
which each cell showed to be cold spot.
}
\description{
Compute a multi-temporal hot and cold spots analysis
employing Getis-Ord Gi* statistic and Mann-Kendall trend test
based on a space-time cube approach.
}
\details{
Compute a multi-temporal hot and cold spots analysis
employing Getis-Ord Gi* statistic and Mann-Kendall trend test
based on a space-time cube approach. If space is defined, a
spatial aggregation (x, y dimensions) by the specified number
of cells is applied. If time is defined, a temporal aggregation
(temporal dimension) by the specified number of observations
(time-steps) is applied. Both of the aforementioned aggregations
are computed based on the statistic specified in stat. Possible
values for stat are "mean" and "max". Default value is "mean".

Center cell is included in the local neighborhood size for Getis
Ord- Gi* and the search limit is the distance specified by dist
employing queen's case. If space is defined, the new raster
resolution after spatial aggregation must be considered when
defining dist.

False Discovery Rate (FDR) correction is applied to potentially
reduce the critical p-value thresholds of the Getis-Ord Gi*
z-scores, in order to account for multiple testing and spatial
dependency. Possible p values to reject the null hypothesis
are 0.1, 0.05, 0.01, 0.001 and 0.0001.

The non-parametric Mann-Kendall test is applied to test for a
monotonic trend based on the Theil-Sen slope modification.
Each cell represents a time-series. x must not have NA values
in valid cells and time-steps should be evenly distributed on
time for optimal results.

Each cell is labeled based on: (1) its statistical significance
in the Getis-Ord Gi* FDR values and Mann-Kendall p-values, (2)
the symbol of the values of the z-scores of GETIS-Ord Gi* (hot,
cold or not significant), (3) the direction of the scores derived
from the Mann-Kendall test (positive, negative or no trend) and
(4) the percentage of time-steps on which cells showed to be hot
or cold spots:

New hot spot: statistically significant hot spot only in the last
time-step.

Increasing hot spot: statistically significant hot spot in >90%
of the time-steps with an increasing trend.

Steady hot spot: statistically significant hot spot in >90% of
the time-steps without trend.

Decreasing hot spot: statistically significant hot spot in >90%
of the time-steps with a decreasing trend.

No pattern: does not meet any of the other criteria.

Decreasing cold spot: statistically significant cold spot in >90%
of the time-steps with a decreasing trend.

Steady cold spot: statistically significant cold spot in >90%
of the time-steps without trend.

Increasing cold spot: statistically significant cold spot in >90%
of the time-steps with an increasing trend.

New cold spot: statistically significant cold spot only in the
last time-step	.
}
\examples{
\dontrun{# For a MODIS LST raster stack
lst_trends <- space_time_trends(x, time = 2, stat = "mean", dist = 2100, p = 0.05)}

}
\references{
Theil, H. (1950). A rank invariant method for linear
and polynomial regression analysis. Nederl. Akad. Wetensch. Proc.
Ser. A 53:386-392 (Part I), 53:521-525 (Part II), 53:1397-1412
(Part III).

Sen, P.K. (1968). Estimates of Regression Coefficient Based on
Kendall's tau. Journal of the American Statistical Association.
63(324):1379-1389.

Kendall, M.G. (1976). Rank Correlation Methods. 4th Ed. Griffin.

Benjamini, Y., and Hochberg, Y. (1995). Controlling
the false discovery rate: a practical and powerful approach
to multiple testing. Journal of the Royal Statistical Society
Series B, 57, 289-300.

Getis, A. and Ord, J. K. (1996). Local spatial statistics:
an overview. In P. Longley and M. Batty (eds) Spatial analysis:
modelling in a GIS environment (Cambridge: Geoinformation
International), 261-277.
}
