% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat_preprocessing.R
\name{toa_ref}
\alias{toa_ref}
\title{toa_ref}
\usage{
toa_ref(x, band = NULL, mult = NULL, add = NULL)
}
\arguments{
\item{x}{Image band in DN to be converted.}

\item{band}{Character. Number of the Landsat band to convert.}

\item{mult}{Reflectance multiplicative band rescaling factor.}

\item{add}{Reflectance additive band rescaling factor.}
}
\value{
Raster layer with TOA reflectance values.
}
\description{
Convert the DN contained in a Landsat band to TOA reflectance.
}
\details{
Convert the DN of a Landsat optical band to TOA reflectance
employing the reflectance multiplicative and additive band rescaling
factors. If band is specified, the function reads the metadata (.txt)
directly from the work directory (folder containing bands as
downloaded from NASA EarthExplorer) and automatically extracts the
multiplicative and additive rescaling factors. These scaling factors
can be manually defined employing mult and add parameters. In this
case, band is ignored.
}
\examples{
# For Landsat 8 band 4 defining band and extracting scaling
factors from metadata
toarefB4 <- toa_ref(B4, band = "4")

# For Landsat 8 band 4 defining manually the multiplicative
and additive scaling factors
toarefB4 <- toa_ref(B4, mult = 0.00033420, add = 0.1)

}
\references{
USGS. (2019). Landsat 8 data users handbook version 4.
USGS Earth Resources Observation and Science (EROS). Sioux Falls,
South Dakota. USA. 106.
}
