% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat_preprocessing.R
\name{toa_refsun}
\alias{toa_refsun}
\title{toa_refsun}
\usage{
toa_refsun(x, band = NULL, mult = NULL, add = NULL, sune = NULL)
}
\arguments{
\item{x}{Image band in DN to be converted.}

\item{band}{Character. Number of the Landsat band to convert.}

\item{mult}{Reflectance multiplicative band rescaling factor.}

\item{add}{Reflectance additive band rescaling factor.}

\item{sune}{Scene center sun elevation angle}
}
\value{
Raster layer with TOA reflectance values corrected for the sun angle.
}
\description{
Convert the DN contained in a Landsat band to TOA reflectance.
}
\details{
Convert the DN of a Landsat optical band to TOA reflectance
employing the reflectance multiplicative and additive band rescaling
factors with a correction for the sun angle. If band is specified,
the function reads the metadata file (.txt) directly from the work
directory (folder containing bands as downloaded from NASA
EarthExplorer) and automatically extracts the multiplicative and
additive rescaling factors and the scene sun elevation angle. These
parameters can be manually defined employing mult, add and sune
parameters. In this case, band is ignored.
}
\examples{
# For Landsat 8 band 4 defining band and extracting scaling
factors from metadata
toarefsunB4 <- toa_refsun(B4, band = "4")

# For Landsat 8 band 4 defining manually the multiplicative and
additive scaling factors and scene sun elevation angle
toarefsunB4 <- toa_refsun (B4, mult = 0.00033420, add = 0.1, sune = 37.22752222)

}
\references{
USGS. (2019). Landsat 8 data users handbook version 4.
USGS Earth Resources Observation and Science (EROS). Sioux Falls,
South Dakota. USA. 106.
}
