% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UHI_indicators.R
\name{uhi_stats}
\alias{uhi_stats}
\title{uhi_stats}
\usage{
uhi_stats(x, y, id = NULL)
}
\arguments{
\item{x}{LST raster layer.}

\item{y}{Raster indicating spatial limits (e.g., land cover).}

\item{id}{Numeric: ID of the spatial limit of the urban area (e.g., urban class).}
}
\value{
A data frame with descriptive statistics and difference
and magnitude UHI indicators.
}
\description{
Computes Urban Heat Island (UHI) indicators based
on a LST raster and spatial limits defined by a raster layer.
}
\details{
Computes LST descriptive statistics (i.e., min, mean,
max and sd) and the difference and magnitude UHI indicators.
Difference is computed as the mean LST urban minus mean LST
other(s) and magnitude is the maximum LST minus mean LST for each
class. x and y must have the same projection and id must be a
numeric value.
}
\examples{
# For urban land use class with id = 2
uhiind <- uhi_stats(lstL8, landuse, id = 2)

}
\references{
Dousset, B. and Gourmelon, F. (2003). Satellite
multi-sensor data analysis of urban surface temperatures and
landcover. ISPRS Journal of Photogrammetry and Remote
Sensing. 58(1-2), 43-54.

Chen, X. L., Zhao, H. M., Li, P. X. and Yin, Z. Y. (2006).
Remote sensing image-based analysis of the relationship
between urban heat island and land use/cover changes. Remote
Sensing of Environment. 104(2), 133-146.

Rajasekar, U. and Weng, Q. H. (2009). Urban heat island monitoring
and analysis using a non-parametric model: A case study of
Indianapolis. ISPRS Journal of Photogrammetry and Remote
Sensing. 64(1), 86-96.

Zhou, J., Li, J. and Yue, J. (2010). Analysis of urban heat
island (UHI) in the Beijing metropolitan area by time-series
MODIS data. IEEE International Geoscience and Remote Sensing
Symposium (IGARSS 2010) (pp. 3327-3330).
}
