% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat_preprocessing.R
\name{untar_landsat}
\alias{untar_landsat}
\title{untar_landsat}
\usage{
untar_landsat(orig, dest = orig)
}
\arguments{
\item{orig}{Path or folder containing the .tar.gz Landsat images.}

\item{dest}{Path or folder to export the uncompressed Landsat images.}
}
\value{
Uncompressed Landsat bands and metadata.
}
\description{
Uncompress .tar.gz folders containing
Landsat images as downloaded from NASA EarthExplorer.
}
\details{
Uncompress .tar.gz files containing Landsat
images and save them in a folder with the original
name of the compressed file. If dest is not specified,
default is the same as orig.
}
\examples{
\dontrun{# Export files to the same folder as input
untar_landsat("C:/Landsat/2017")

# Export files to a folder different than the input
untar_landsat("C:/Landsat/2017", "C:/Landsat")}
}
\references{
https://earthexplorer.usgs.gov/
}
