% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat_processing.R
\name{urbi}
\alias{urbi}
\title{urbi}
\usage{
urbi(x, y)
}
\arguments{
\item{x}{SWIR band.}

\item{y}{NIR band.}
}
\value{
UI values ranging from -1 to 1.
}
\description{
Compute the Urban Index (UI).
}
\details{
Compute the Urban Index (UI) employing the surface
reflectance SWIR and NIR bands of a satellite image.
}
\examples{
\dontrun{# For Landsat 8 bands
urbind <- urbi (B7, B5)}

}
\references{
Kawamura, M., Jayamana, S. and Tsujiko, Y. (1996).
Relation between social and environmental conditions in
Colombo Sri Lanka and the urban index estimated by satellite
remote sensing data. Int. Arch. Photogramm. Remote Sens.
31(Part B7), 321-326.
}
